+++ src/Providers/WidgetServiceProvider.php
@@ -32,6 +32,7 @@ class WidgetServiceProvider extends ServiceProvider
         AppShellWidgets::ENUM_ICON => Widgets\EnumIcon::class,
         AppShellWidgets::FILTER => Widgets\Filter::class,
         AppShellWidgets::FILTER_SET => Widgets\FilterSet::class,
+        AppShellWidgets::IMAGE => Widgets\Image::class,
         AppShellWidgets::RAW_HTML => Widgets\RawHtml::class,
         AppShellWidgets::SHOW_DATE => Widgets\ShowDate::class,
         AppShellWidgets::SHOW_DATETIME => Widgets\ShowDateTime::class,