+++ src/Widgets/Badge.php
@@ -55,7 +55,7 @@ class Badge implements Widget
         $colorDef = $options['color'] ?? null;
         $bgColor = $this->parseColorDefinition(
             is_string($colorDef) ? $this->resolveSubstitutions($colorDef, $data) : $colorDef,
-            $value
+            $value, null, false, $data
         );
         $contextClass = $bgColor->themeColor->isNone() ? '' : 'text-bg-' . $bgColor->themeColor->value();
         $result['class'] = "badge rounded-pill $contextClass " . ($options['class'] ?? '');