+++ src/resources/views/widgets/image.blade.php
@@ -1,0 +1,1 @@
+<img src="{{ $src }}" @if($class)class="{{ $class }}"@endif @if($style)style="{{ $style }}"@endif @if($tooltip)title="{{ $tooltip }}"@endif>